requirements.txt
click==8.1.7
Jinja2==3.1.4
PyYAML==6.0.2
python-dotenv==1.0.1
requests==2.32.3

pytest
